+++ assets/javascripts/kitten/components/structure/cards/content-card/index.js
@@ -50,6 +50,11 @@ const StyledCard = styled.article`
     }
   }
 
+  .k-ContentCard__content--tag {
+    display: flex;
+    gap: ${pxToRem(10)};
+  }
+
   .k-ContentCard__content--fixedHeight {
     max-height: ${pxToRem(150)};
     overflow: hidden;